requirements.txt
pip>=25.2
openai>=1.104.2
python-dotenv
pinecone
langchain_openai
langchain
langchain-community
lark